Create the Code
The code first imports sqlite3 and then it makes a connection using file ‘15.2_Counting_Email_in_a_Database.sqlite’. If the table in the file already exists, it removes the table to start over/start clean. 

Next, it creates a new table with an ‘email’ column and a ‘count’ column. Then, it prompts the user to give a file name, and defaults to ‘mbox-short.txt’ if this is left blank. It opens and reads the file.

Then, the code loops through the file, looking for lines that start with ‘From:’ and skipping the rest. It splits these lines and returns the 2nd segment (the email) and then splits the email at the ‘@’ character, returning the 2nd segment, the email domain only (indicating the organization).

Next, the code opens the record, selects the ‘count’ column from the ‘Counts’ table, and uses ‘email_pieces’ for the ‘email’ column. Then it grabs the next row from the database. If the email domain (organization) does not have a row, it creates a row, adds the email domain under ‘email’, and adds a count of 1 for ‘count’. If the email domain (organization) has a row, it updates, adding 1 to the count. Then it commits (writes to memory).
 
Finally, the code selects all the ‘emails’ and corresponding ‘counts’ (in the ‘Counts’ table), sorts by ‘counts’ in descending order, executes the ‘sqlstr’ statement, prints the emails and the corresponding counts, and closes the connection.